rename data_generator to prefix_data_generator
