+++ benchmarks/prefix_data_generator/prefix_analyzer.py
@@ -16,7 +16,7 @@
 import json
 from collections import Counter
 
-from data_generator.logging_utils import calculate_and_print_statistics
+from prefix_data_generator.logging_utils import calculate_and_print_statistics
 
 
 class PrefixAnalyzer:
