+++ benchmarks/pyproject.toml
@@ -49,7 +49,7 @@ dependencies = [
 ]
 
 [project.scripts]
-datagen = "data_generator.cli:main"
+datagen = "prefix_data_generator.cli:main"
 
 [project.urls]
 Repository = "https://github.com/ai-dynamo/dynamo.git"
@@ -59,10 +59,10 @@ requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["data_generator"]
+packages = ["prefix_data_generator"]
 
 [tool.setuptools.package-data]
-data_generator = ["**/*.py"]
+prefix_data_generator = ["**/*.py"]
 
 [tool.mypy]
 explicit_package_bases = true
